Rename spatial navigation folder
