+++ src/ts/spatialnavigation/gethtmlelementsfromcomponents.ts
