+++ src/ts/spatialnavigation/navigationalgorithm.ts
