+++ src/ts/spatialnavigation/nodeeventsubscriber.ts
