+++ src/ts/spatialnavigation/seekbarhandler.ts
