+++ src/ts/spatialnavigation/spatialnavigation.ts
