+++ src/ts/spatialnavigation/typeguards.ts
